reorganise heading components
